requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
boto3
azure-identity
azure-mgmt-costmanagement
azure-mgmt-compute
google-cloud-billing
google-cloud-monitoring
requests
pandas
numpy
scikit-learn
streamlit
plotly
python-dotenv
schedule